WIP unravel stack index
